fix: 修复日期展示 fromNow 展示为德文的问题 Close: #6914 (#6926)
